requirements.txt
requests
pydash
Pillow